requirements.txt
flask
flask-restful
flask-cors
mysql-connector-python